Add HAProxy into the docker-compose mix
Results of putting HAProxy in front of the current ProxyMXDev (which has 32 cores and 32GB of RAM).
Tested with unencrypted mails since we pretty much know about core's "speed of light" of
slightly below 20msgs per second per halfway decent CPU core.

1 instance: 16.73s to hand-off 1000 messages into the Postfix queue, 5mins + 27s until fully delivered on the final mailbox server.
2 instances: 16.22s to hand-off, 2mins + 26s to fully deliver
4 instances: 20.43s and 1min + 21s
16 instances: 22.58s and 42s

Congestion hits beyond this point so 16 is our sweet-spot on this machine.

24 instances: 29.65s and 48s
32 instances: 32s and 55s

PS: using Postfix as a load-balancer was pretty much a dead-end. It worked somewhat
but it introduces all kind of more or less expected side-effects